renamed internal session class to InternalSession
