+++ addon/internal-session.js
