+++ tests/unit/initializers/setup-session-test.js
@@ -4,7 +4,7 @@ import { describe, beforeEach } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import setupSession from 'ember-simple-auth/initializers/setup-session';
-import Session from 'ember-simple-auth/session';
+import InternalSession from 'ember-simple-auth/internal-session';
 import Configuration from 'ember-simple-auth/configuration';
 
 describe('setupSession', () => {
@@ -21,7 +21,7 @@ describe('setupSession', () => {
     sinon.spy(registry, 'register');
     setupSession(registry);
 
-    expect(registry.register).to.have.been.calledWith('session:main', Session);
+    expect(registry.register).to.have.been.calledWith('session:main', InternalSession);
   });
 
   it('injects the session store into the session', () => {
