+++ tests/unit/mixins/unauthenticated-route-mixin-test.js
@@ -5,7 +5,7 @@ import { describe, beforeEach } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import UnauthenticatedRouteMixin from 'ember-simple-auth/mixins/unauthenticated-route-mixin';
-import Session from 'ember-simple-auth/session';
+import InternalSession from 'ember-simple-auth/internal-session';
 import Configuration from 'ember-simple-auth/configuration';
 import EphemeralStore from 'ember-simple-auth/stores/ephemeral';
 
@@ -16,7 +16,7 @@ describe('UnauthenticatedRouteMixin', () => {
 
   describe('#beforeModel', () => {
     beforeEach(() => {
-      session    = Session.create({ store: EphemeralStore.create() });
+      session    = InternalSession.create({ store: EphemeralStore.create() });
       transition = {
         abort() {},
         send() {}
